refactor(toolbar_rich_text_test.dart): move LinkMenu test to its own file
move LinkMenu test according to convention - test is at same folder structure, as tested code
